# docs/04-Backend/archive/v3-openAPI.yaml
openapi: "3.1.0" 
info:
  title: "MWAP API"
  version: "1.0.0"
  description: "Full OpenAPI spec for the MWAP backend aligned with the domain model, contract, and architecture."
servers:
  - url: "https://api.mwap.dev/api/v1"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /openapi:
    get:
      summary: Get complete OpenAPI specification
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [json, yaml]
            default: json
        - in: query
          name: includeExamples
          schema:
            type: boolean
            default: false
        - in: query
          name: minify
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: OpenAPI specification
          content:
            application/json:
              schema:
                type: object
            application/x-yaml:
              schema:
                type: string
  /openapi/info:
    get:
      summary: Get OpenAPI information summary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /openapi/validate:
    get:
      summary: Validate OpenAPI specification
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Validation successful
          content:
            application/json:
              schema:
                type: object
...
  /openapi/cache/status:
    get:
      summary: Get OpenAPI cache status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /openapi/cache/invalidate:
    post:
      summary: Invalidate OpenAPI cache (admin only)
      description: Invalidates the OpenAPI cache; requires superadmin role.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cache invalidated
          content:
            application/json:
              schema:
                type: object
  /openapi/validation/history:
    get:
      summary: Get OpenAPI validation history
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /openapi/validation/ci-report:
    get:
      summary: Generate CI/CD validation report
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [json, junit, text]
            default: text
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                type: object
            application/xml:
              schema:
                type: string
  /openapi/validation/monitor:
    post:
      summary: Monitor OpenAPI validation
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
...
  /openapi/performance/benchmark:
    post:
      summary: Run OpenAPI performance benchmarks
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                iterations:
                  type: integer
                  default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /openapi/performance/optimize-cache:
    post:
      summary: Optimize OpenAPI cache configuration
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /openapi/security/audit:
    post:
      summary: Perform OpenAPI security audit
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /openapi/security/sanitized:
    get:
      summary: Get sanitized OpenAPI specification
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /openapi/security/audit-log:
    get:
      summary: Get OpenAPI security audit log
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object